pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-summary-tool"
version = "0.1.0"
description = "Local Telegram export summarizer CLI"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "PyYAML>=6.0.1",
    "pycryptodome>=3.20",
    "mmh3>=4.0,<5.0",
    "fastapi>=0.100",
    "uvicorn>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
tg-summary = "telegram_summary_tool.cli:main"
tg-web = "telegram_summary_tool.web:main"
tg-doctor = "telegram_summary_tool.doctor:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
